+++ core/src/main/kotlin/net/corda/core/flows/NotaryFlow.kt
@@ -8,7 +8,9 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.SignedData
 import net.corda.core.crypto.keys
 import net.corda.core.identity.Party
-import net.corda.core.node.services.*
+import net.corda.core.node.services.NotaryService
+import net.corda.core.node.services.TrustedAuthorityNotaryService
+import net.corda.core.node.services.UniquenessProvider
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.ProgressTracker
@@ -56,14 +58,13 @@ object NotaryFlow {
                 throw NotaryException(NotaryError.TransactionInvalid(ex))
             }
 
-            val payload: Any = if (serviceHub.networkMapCache.isValidatingNotary(notaryParty)) {
-                stx
-            } else {
-                wtx.buildFilteredTransaction(Predicate { it is StateRef || it is TimeWindow })
-            }
-
             val response = try {
-                sendAndReceiveWithRetry<List<DigitalSignature.WithKey>>(notaryParty, payload)
+                if (serviceHub.networkMapCache.isValidatingNotary(notaryParty)) {
+                    subFlow(SendTransactionWithRetry(notaryParty, stx))
+                    receive<List<DigitalSignature.WithKey>>(notaryParty)
+                } else {
+                    sendAndReceiveWithRetry(notaryParty, wtx.buildFilteredTransaction(Predicate { it is StateRef || it is TimeWindow }))
+                }
             } catch (e: NotaryException) {
                 if (e.error is NotaryError.Conflict) {
                     e.error.conflict.verified()
@@ -138,4 +139,13 @@ sealed class NotaryError {
     data class TransactionInvalid(val cause: Throwable) : NotaryError() {
         override fun toString() = cause.toString()
     }
+}
+
+/**
+ * The [SendTransactionWithRetry] flow is equivalent to [SendTransactionFlow] but using [sendAndReceiveWithRetry]
+ * instead of [sendAndReceive], [SendTransactionWithRetry] is intended to be use by the notary client only.
+ */
+private class SendTransactionWithRetry(otherSide: Party, stx: SignedTransaction) : SendTransactionFlow(otherSide, stx) {
+    @Suspendable
+    override fun sendPayloadAndReceiveDataRequest(payload: Any?) = sendAndReceiveWithRetry<FetchDataFlow.Request>(otherSide, payload!!)
 }