+++ core/src/main/kotlin/net/corda/core/flows/ResolveTransactionsFlow.kt
@@ -1,7 +1,6 @@
 package net.corda.core.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.checkedAdd
 import net.corda.core.crypto.SecureHash
 import net.corda.core.getOrThrow
 import net.corda.core.identity.Party
@@ -9,13 +8,16 @@ import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
+import net.corda.core.utilities.UntrustworthyData
+import net.corda.core.utilities.unwrap
 import java.util.*
-
-// TODO: This code is currently unit tested by TwoPartyTradeFlowTests, it should have its own tests.
+import kotlin.collections.LinkedHashSet
 
 // TODO: It may be a clearer API if we make the primary c'tor private here, and only allow a single tx to be "resolved".
 
 /**
+ * The [ResolveTransactionsFlow] corresponds to the [SendTransactionFlow].
+ *
  * This flow is used to verify the validity of a transaction by recursively checking the validity of all the
  * dependencies. Once a transaction is checked it's inserted into local storage so it can be relayed and won't be
  * checked again.
@@ -28,24 +30,21 @@ import java.util.*
  *
  * The flow returns a list of verified [LedgerTransaction] objects, in a depth-first order.
  */
-class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
-                              private val otherSide: Party) : FlowLogic<List<LedgerTransaction>>() {
+class ResolveTransactionsFlow(private val otherSide: Party, private val transactionData: ResolvableTransactionData,
+                              private val verifySignatures: Boolean = true, private val verifyTransaction: Boolean = true) : FlowLogic<List<LedgerTransaction>>() {
+    constructor(otherSide: Party, transactionData: ResolvableTransactionData) : this(otherSide, transactionData, true, true)
 
     companion object {
         private fun dependencyIDs(wtx: WireTransaction) = wtx.inputs.map { it.txhash }.toSet()
-
         /**
          * Topologically sorts the given transactions such that dependencies are listed before dependers. */
         @JvmStatic
         fun topologicalSort(transactions: Collection<SignedTransaction>): List<SignedTransaction> {
             // Construct txhash -> dependent-txs map
-            val forwardGraph = HashMap<SecureHash, HashSet<SignedTransaction>>()
-            transactions.forEach { stx ->
-                stx.tx.inputs.forEach { (txhash) ->
+            val forwardGraph = transactions.flatMap { stx -> stx.tx.inputs.map { it.txhash to stx } }
+                    .groupBy { it.first }
                     // Note that we use a LinkedHashSet here to make the traversal deterministic (as long as the input list is)
-                    forwardGraph.getOrPut(txhash) { LinkedHashSet() }.add(stx)
-                }
-            }
+                    .mapValues { LinkedHashSet(it.value.map { it.second }) }
 
             val visited = HashSet<SecureHash>(transactions.size)
             val result = ArrayList<SignedTransaction>(transactions.size)
@@ -57,81 +56,60 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
                     result.add(transaction)
                 }
             }
-
             transactions.forEach(::visit)
-
             result.reverse()
             require(result.size == transactions.size)
             return result
         }
-
     }
 
     @CordaSerializable
     class ExcessivelyLargeTransactionGraph : Exception()
 
-    // Transactions to verify after the dependencies.
-    private var stx: SignedTransaction? = null
-    private var wtx: WireTransaction? = null
-
     // TODO: Figure out a more appropriate DOS limit here, 5000 is simply a very bad guess.
     /** The maximum number of transactions this flow will try to download before bailing out. */
     var transactionCountLimit = 5000
 
-    /**
-     * Resolve the full history of a transaction and verify it with its dependencies.
-     */
-    constructor(stx: SignedTransaction, otherSide: Party) : this(stx.tx, otherSide) {
-        this.stx = stx
-    }
-
-    /**
-     * Resolve the full history of a transaction and verify it with its dependencies.
-     */
-    constructor(wtx: WireTransaction, otherSide: Party) : this(dependencyIDs(wtx), otherSide) {
-        this.wtx = wtx
-    }
-
     @Suspendable
     @Throws(FetchDataFlow.HashNotFound::class)
     override fun call(): List<LedgerTransaction> {
-        val newTxns: Iterable<SignedTransaction> = topologicalSort(downloadDependencies(txHashes))
-
-        // For each transaction, verify it and insert it into the database. As we are iterating over them in a
-        // depth-first order, we should not encounter any verification failures due to missing data. If we fail
-        // half way through, it's no big deal, although it might result in us attempting to re-download data
-        // redundantly next time we attempt verification.
-        val result = ArrayList<LedgerTransaction>()
-
-        for (stx in newTxns) {
-            // Resolve to a LedgerTransaction and then run all contracts.
-            val ltx = stx.toLedgerTransaction(serviceHub)
-            // Block on each verification request.
-            // TODO We could recover some parallelism from the dependency graph.
-            serviceHub.transactionVerifierService.verify(ltx).getOrThrow()
-            serviceHub.recordTransactions(stx)
-            result += ltx
+        val stx = when (transactionData) {
+            is SignedTransaction -> transactionData
+            is ResolvableTransactionData.Transaction -> transactionData.stx
+            else -> null
         }
 
-        // If this flow is resolving a specific transaction, make sure we have its attachments and then verify
-        // it as well, but don't insert to the database. Note that when we were given a SignedTransaction (stx != null)
-        // we *could* insert, because successful verification implies we have everything we need here, and it might
-        // be a clearer API if we do that. But for consistency with the other c'tor we currently do not.
-        //
-        // If 'stx' is set, then 'wtx' is the contents (from the c'tor).
-        val wtx = stx?.verifyRequiredSignatures() ?: wtx
-        wtx?.let {
-            fetchMissingAttachments(listOf(it))
+        val wtx = if (verifySignatures) {
+            stx?.verifyRequiredSignatures()
+        } else {
+            stx?.tx
+        }
+
+        // Start fetching data.
+        val newTxns = downloadDependencies(transactionData.dependencies)
+        fetchMissingAttachments(newTxns.flatMap { it.tx.attachments } + (wtx?.attachments ?: emptyList()))
+        send(otherSide, FetchDataFlow.EndRequest)
+        // Finish fetching data.
+
+        val result = topologicalSort(newTxns).map {
+            // For each transaction, verify it and insert it into the database. As we are iterating over them in a
+            // depth-first order, we should not encounter any verification failures due to missing data. If we fail
+            // half way through, it's no big deal, although it might result in us attempting to re-download data
+            // redundantly next time we attempt verification.
             val ltx = it.toLedgerTransaction(serviceHub)
-            ltx.verify()
-            result += ltx
+            serviceHub.transactionVerifierService.verify(ltx).getOrThrow()
+            serviceHub.recordTransactions(it)
+            ltx
         }
-
-        return result
+        return if (wtx == null || !verifyTransaction) {
+            result
+        } else {
+            result + wtx.toLedgerTransaction(serviceHub).apply { verify() }
+        }
     }
 
     @Suspendable
-    private fun downloadDependencies(depsToCheck: Set<SecureHash>): Collection<SignedTransaction> {
+    private fun downloadDependencies(depsToCheck: Set<SecureHash>): List<SignedTransaction> {
         // Maintain a work queue of all hashes to load/download, initialised with our starting set. Then do a breadth
         // first traversal across the dependency graph.
         //
@@ -152,36 +130,32 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
         nextRequests.addAll(depsToCheck)
         val resultQ = LinkedHashMap<SecureHash, SignedTransaction>()
 
-        val limit = transactionCountLimit
-        check(limit > 0) { "$limit is not a valid count limit" }
-        var limitCounter = 0
+        var limitCounter = transactionCountLimit
+        check(limitCounter > 0) { "$limitCounter is not a valid count limit" }
+
         while (nextRequests.isNotEmpty()) {
+            if (limitCounter < 0) throw ExcessivelyLargeTransactionGraph()
+            limitCounter -= nextRequests.size
+
             // Don't re-download the same tx when we haven't verified it yet but it's referenced multiple times in the
             // graph we're traversing.
             val notAlreadyFetched = nextRequests.filterNot { it in resultQ }.toSet()
-            nextRequests.clear()
-
-            if (notAlreadyFetched.isEmpty())   // Done early.
-                break
+            if (notAlreadyFetched.isEmpty())
+                break // Done early.
 
             // Request the standalone transaction data (which may refer to things we don't yet have).
-            val downloads: List<SignedTransaction> = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
-
-            fetchMissingAttachments(downloads.map { it.tx })
+            val downloads = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
 
             for (stx in downloads)
                 check(resultQ.putIfAbsent(stx.id, stx) == null)   // Assert checks the filter at the start.
 
             // Add all input states to the work queue.
             val inputHashes = downloads.flatMap { it.tx.inputs }.map { it.txhash }
+
+            nextRequests.clear()
             nextRequests.addAll(inputHashes)
-
-            limitCounter = limitCounter checkedAdd nextRequests.size
-            if (limitCounter > limit)
-                throw ExcessivelyLargeTransactionGraph()
         }
-
-        return resultQ.values
+        return resultQ.values.toList()
     }
 
     /**
@@ -189,12 +163,40 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
      * first in the returned list and thus doesn't have any unverified dependencies.
      */
     @Suspendable
-    private fun fetchMissingAttachments(downloads: List<WireTransaction>) {
+    private fun fetchMissingAttachments(downloads: List<SecureHash>) {
         // TODO: This could be done in parallel with other fetches for extra speed.
-        val missingAttachments = downloads.flatMap { wtx ->
-            wtx.attachments.filter { serviceHub.attachments.openAttachment(it) == null }
-        }
+        val missingAttachments = downloads.filter { serviceHub.attachments.openAttachment(it) == null }
         if (missingAttachments.isNotEmpty())
             subFlow(FetchAttachmentsFlow(missingAttachments.toSet(), otherSide))
     }
+}
+
+/**
+ * Receive and resolve transaction
+ */
+@Suspendable
+fun <T : ResolvableTransactionData> FlowLogic<*>.receiveTransaction(clazz: Class<T>, otherSide: Party,
+                                                                    verifySignatures: Boolean = true,
+                                                                    verifyTransaction: Boolean = true): UntrustworthyData<T> {
+    return receive(clazz, otherSide).unwrap {
+        subFlow(ResolveTransactionsFlow(otherSide, it, verifySignatures, verifyTransaction))
+        UntrustworthyData(it)
+    }
+}
+
+inline fun <reified T : ResolvableTransactionData> FlowLogic<*>.receiveTransaction(otherSide: Party,
+                                                                                   verifySignature: Boolean = true,
+                                                                                   verifyTransaction: Boolean = true)
+        = receiveTransaction(T::class.java, otherSide, verifySignature, verifyTransaction)
+
+@CordaSerializable
+interface ResolvableTransactionData {
+    val dependencies: Set<SecureHash>
+
+    interface Transaction : ResolvableTransactionData {
+        override val dependencies: Set<SecureHash> get() = stx.dependencies
+        val stx: SignedTransaction
+    }
+
+    data class TransactionHashes(override val dependencies: Set<SecureHash>) : ResolvableTransactionData
 }