+++ docs/source/example-code/src/main/kotlin/net/corda/docs/FlowCookbook.kt
@@ -382,12 +382,12 @@ object FlowCookbook {
             // dependencies, we'd need to download all of these dependencies
             // using``ResolveTransactionsFlow`` before verifying it.
             // DOCSTART 13
-            subFlow(ResolveTransactionsFlow(twiceSignedTx, counterparty))
+          //  subFlow(ResolveTransactionsFlow(twiceSignedTx, counterparty))
             // DOCEND 13
 
             // We can also resolve a `StateRef` dependency chain.
             // DOCSTART 14
-            subFlow(ResolveTransactionsFlow(setOf(ourStateRef.txhash), counterparty))
+          //  subFlow(ResolveTransactionsFlow(setOf(ourStateRef.txhash), counterparty))
             // DOCEND 14
 
             // A ``SignedTransaction`` is a pairing of a ``WireTransaction``