+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryFlow.kt
@@ -6,7 +6,6 @@ import net.corda.core.flows.*
 import net.corda.core.identity.Party
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.transactions.SignedTransaction
-import net.corda.core.transactions.WireTransaction
 import net.corda.core.utilities.unwrap
 import java.security.SignatureException
 
@@ -24,26 +23,13 @@ class ValidatingNotaryFlow(otherSide: Party, service: TrustedAuthorityNotaryServ
      */
     @Suspendable
     override fun receiveAndVerifyTx(): TransactionParts {
-        val stx = receive<SignedTransaction>(otherSide).unwrap { it }
-        checkSignatures(stx)
-        val wtx = stx.tx
-        validateTransaction(wtx)
-        return TransactionParts(wtx.id, wtx.inputs, wtx.timeWindow)
-    }
-
-    private fun checkSignatures(stx: SignedTransaction) {
         try {
-            stx.verifySignaturesExcept(serviceHub.myInfo.notaryIdentity.owningKey)
-        } catch(e: SignatureException) {
-            throw NotaryException(NotaryError.TransactionInvalid(e))
-        }
-    }
-
-    @Suspendable
-    fun validateTransaction(wtx: WireTransaction) {
-        try {
-            resolveTransaction(wtx)
-            wtx.toLedgerTransaction(serviceHub).verify()
+            return receiveTransaction<SignedTransaction>(otherSide, verifySignature = false).unwrap {
+                it.verifySignaturesExcept(serviceHub.myInfo.notaryIdentity.owningKey)
+                val wtx = it.tx
+                wtx.toLedgerTransaction(serviceHub).verify()
+                TransactionParts(wtx.id, wtx.inputs, wtx.timeWindow)
+            }
         } catch (e: Exception) {
             throw when (e) {
                 is TransactionVerificationException,
@@ -52,7 +38,4 @@ class ValidatingNotaryFlow(otherSide: Party, service: TrustedAuthorityNotaryServ
             }
         }
     }
-
-    @Suspendable
-    private fun resolveTransaction(wtx: WireTransaction) = subFlow(ResolveTransactionsFlow(wtx, otherSide))
 }